requirements.txt
# Django e REST framework
Django==4.2.7
djangorestframework==3.14.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
drf-spectacular==0.26.4
django-redis==5.3.0
django-filter==23.5

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.7

# Static files
whitenoise[brotli]==6.6.0

# CORS per API
django-cors-headers==4.3.1

# Utilities
Pillow==10.1.0
python-decouple==3.8

# WebSocket e canali
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Storage e Media
cloudinary==1.36.0
django-cloudinary-storage==0.3.0

# Server WSGI
gunicorn==21.2.0

# Security
django-cleanup==8.0.0
python-magic==0.4.27

# Production
gunicorn==21.2.0

# Development (optional)
django-debug-toolbar==4.2.0

django-filter==23.5 # Per filtri avanzati nelle API
channels==4.0.0  # Per WebSocket
channels-redis==4.1.0  # Per WebSocket con Redis
